# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /app/auth/challenge:
    post:
      operationId: createWalletChallenge
      tags: [bot]
      summary: Create a single-use wallet ownership challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletChallengeInput"
      responses:
        "200":
          description: Wallet challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletChallenge"
  /app/auth/verify:
    post:
      operationId: verifyWalletChallenge
      tags: [bot]
      summary: Verify wallet ownership and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletVerificationInput"
      responses:
        "200":
          description: Authenticated app state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserBotDashboard"
...
  /app/bot/setup:
    post:
      operationId: setupUserBot
      tags: [bot]
      summary: Validate a KCC20 token and create an isolated bot wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotSetupInput"
      responses:
        "200":
          description: Updated user bot dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserBotDashboard"
  /app/bot/activation/verify:
    post:
      operationId: verifyBotActivation
      tags: [bot]
      summary: Verify the 100 KAS activation payment on-chain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationVerificationInput"
      responses:
        "200":
          description: Updated user bot dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserBotDashboard"
  /app/bot/covenant/change:
    post:
      operationId: changeUserBotCovenant
      tags: [bot]
      summary: Change the covenant after all positions are sold and require a new activation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CovenantChangeInput"
      responses:
        "200":
          description: Updated user bot dashboard awaiting activation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserBotDashboard"
...
  /app/bot/withdraw:
    post:
      operationId: prepareUserBotKasWithdrawal
      tags: [bot]
      summary: Prepare a bot-wallet withdrawal for connected-wallet fee signing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KasWithdrawalInput"
      responses:
        "200":
          description: Withdrawal transaction awaiting the connected wallet signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KasWithdrawalPreparation"
  /app/bot/withdraw/submit:
    post:
      operationId: submitUserBotKasWithdrawal
      tags: [bot]
      summary: Validate and submit a connected-wallet-signed withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KasWithdrawalSubmission"
      responses:
        "200":
          description: Submitted KAS withdrawal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KasWithdrawalResult"
components:
  schemas:
...
    WalletChallengeInput:
      type: object
      required: [walletAddress, publicKey]
      properties:
        walletAddress:
          type: string
          minLength: 20
        publicKey:
          type: string
          minLength: 64
    WalletChallenge:
      type: object
      required: [challengeId, message, expiresAt]
      properties:
        challengeId:
          type: string
        message:
          type: string
        expiresAt:
          type: string
          format: date-time
    WalletVerificationInput:
      allOf:
        - $ref: "#/components/schemas/WalletChallengeInput"
        - type: object
          required: [challengeId, signature]
          properties:
            challengeId:
              type: string
            signature:
              type: string
              minLength: 64
    BotSetupInput:
      type: object
      required: [tokenId]
      properties:
        tokenId:
          type: string
          pattern: "^[a-fA-F0-9]{64}$"
    CovenantChangeInput:
      type: object
      required: [tokenId]
      properties:
        tokenId:
          type: string
          pattern: "^[a-fA-F0-9]{64}$"
    KasWithdrawalInput:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
          enum: [amount, max]
        amountKas:
          type: string
          pattern: "^(?:0|[1-9][0-9]*)(?:\\.[0-9]{1,8})?$"
    KasWithdrawalPreparation:
      type: object
      required: [txJsonString, signInputs, destinationAddress, amountKas, feeKas, remainingBalanceKas]
      properties:
        txJsonString:
          type: string
        approvalMessage:
          type: string
          description: Fresh wallet approval message for bot-funded Max withdrawals.
        signInputs:
          type: array
          items:
            type: object
            required: [index, sighashType]
            properties:
              index:
                type: integer
              sighashType:
                type: integer
        destinationAddress:
          type: string
        amountKas:
          type: string
        feeKas:
          type: string
        remainingBalanceKas:
          type: string
    KasWithdrawalSubmission:
      type: object
      required: [signedTransaction]
      properties:
        signedTransaction:
          type: string
    KasWithdrawalResult:
      type: object
      required: [transactionId, destinationAddress, amountKas, feeKas, remainingBalanceKas]
      properties:
        transactionId:
          type: string
        destinationAddress:
          type: string
        amountKas:
          type: string
        feeKas:
          type: string
        remainingBalanceKas:
          type: string
...
    ActivationVerificationInput:
      type: object
      required: [transactionId]
      properties:
        transactionId:
          type: string
          pattern: "^[a-fA-F0-9]{64}$"
    FixedStrategy:
      type: object
      required: [orderSizeKas, buyCount, sellCount, tradesPerHour, activationFeeKas, activationAddress]
      properties:
        orderSizeKas:
          type: number
        buyCount:
          type: integer
        sellCount:
          type: integer
        tradesPerHour:
          type: integer
        activationFeeKas:
          type: number
        activationAddress:
          type: string
    UserBotDashboard:
      type: object
      required: [authenticated, walletAddress, strategy, bot]
      properties:
        authenticated:
          type: boolean
        walletAddress:
          type: string
        strategy:
          $ref: "#/components/schemas/FixedStrategy"
        bot:
          type: ["object", "null"]
          properties:
            id:
              type: string
            botAddress:
              type: string
            tokenId:
              type: ["string", "null"]
            tokenSymbol:
              type: ["string", "null"]
            configurationVersion:
              type: integer
            activationPaid:
              type: boolean
            activationTxId:
              type: ["string", "null"]
            status:
              type: string
              enum: [setup, ready, running, paused, stopped]
            phase:
              type: string
              enum: [buying, selling]
            completedBuys:
              type: integer
            completedSells:
              type: integer
            totalTrades:
              type: integer
            nextRunAt:
              type: ["string", "null"]
              format: date-time
            lastTradeAt:
              type: ["string", "null"]
              format: date-time
            stopReason:
              type: ["string", "null"]
            managedTokenAmount:
              type: string
            botKasBalance:
              type: number
            tradeHistory:
              type: array
              items:
                type: object
                properties:
                  action:
                    type: string
                    enum: [buy, sell]
                  transactionId:
                    type: string
                  executedAt:
                    type: string
                    format: date-time
                  tokenAmount:
                    type: string
                  tokenId:
                    type: ["string", "null"]
                  tokenSymbol:
                    type: ["string", "null"]
                required: [action, transactionId, executedAt, tokenAmount, tokenId, tokenSymbol]
          required: [id, botAddress, tokenId, tokenSymbol, configurationVersion, activationPaid, activationTxId, status, phase, completedBuys, completedSells, totalTrades, nextRunAt, lastTradeAt, stopReason, managedTokenAmount, botKasBalance, tradeHistory]